tests/wpt/tests/svg/text/inheritance.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="8000px">
  <title>Inheritance of text properties</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/text.html"/>
    <h:meta name="assert" content="text properties do not inherit."/>
    <h:meta name="assert" content="text properties have expected initial values."/>
  </metadata>
  <g id="container">
      <g id="target"></g>
  </g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/inheritance-testcommon.js"/>
  <script><![CDATA[

assert_not_inherited('inline-size', 'auto', '10px');
assert_not_inherited('shape-inside', 'auto', 'url("https://example/com/")');
assert_not_inherited('shape-subtract', 'auto', 'url("https://example/com/")');
assert_not_inherited('shape-margin', '0px', '10px');
assert_inherited('text-anchor', 'start', 'middle');
assert_not_inherited('text-decoration-fill', 'none', 'url("https://example/com/")');
assert_not_inherited('text-decoration-stroke', 'none', 'url("https://example/com/")');

  ]]></script>
</svg>
